requirements.txt
streamlit
pywebview